# optimize serialization functions