benchmark data:
```bash
go test -test.bench=".*" -benchmem
```

after optimize:
```
BenchmarkWriteVarUint-4         50000000                26.9 ns/op            16 B/op          1 allocs/op
BenchmarkWriteVarBytes-4        30000000                49.9 ns/op            25 B/op          1 allocs/op
BenchmarkWriteVarString-4       20000000                74.3 ns/op            37 B/op          2 allocs/op
BenchmarkReadVarUint-4          20000000                75.9 ns/op            64 B/op          2 allocs/op
BenchmarkReadVarBytes-4         20000000               105 ns/op              80 B/op          3 allocs/op
BenchmarkReadVarString-4        10000000               147 ns/op              96 B/op          4 allocs/op
BenchmarkSerialize-4             2000000               764 ns/op             384 B/op         20 allocs/op
```
original ：
```
BenchmarkWriteVarUint-4         10000000               143 ns/op             128 B/op          3 allocs/op
BenchmarkWriteVarBytes-4        10000000               157 ns/op             141 B/op          3 allocs/op
BenchmarkWriteVarString-4       10000000               177 ns/op             149 B/op          4 allocs/op
BenchmarkReadVarUint-4           5000000               242 ns/op             192 B/op          6 allocs/op
BenchmarkReadVarBytes-4          5000000               242 ns/op             192 B/op          6 allocs/op
BenchmarkReadVarString-4         5000000               282 ns/op             208 B/op          7 allocs/op
BenchmarkSerialize-4              500000              2989 ns/op            2016 B/op         66 allocs/op

```

Signed-off-by: laizy <laizhichao@onchain.com>